drivers:cc110x_ng: fix header too

+++ drivers/include/cc110x_ng/cc110x-interface.h
@@ -131,9 +131,9 @@ void cc110x_wakeup_from_rx(void);
 void cc110x_switch_to_pwd(void);
 
 void cc110x_disable_interrupts(void);
-int16_t cc110x_set_config_channel(uint8_t channr);
-int16_t cc110x_set_channel(uint8_t channr);
-int16_t cc110x_get_channel(void);
+int32_t cc110x_set_config_channel(uint8_t channr);
+int32_t cc110x_set_channel(uint8_t channr);
+int32_t cc110x_get_channel(void);
 
 radio_address_t cc110x_set_address(radio_address_t addr);
 radio_address_t cc110x_set_config_address(radio_address_t addr);
